infra/containers/databento_ingestor/requirements.txt
databento>=0.15.0
azure-eventhub>=5.11.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.7.0
